requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pinecone-client==2.2.4
openai==1.3.7
pypdf2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
pydantic>=2.5.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 